feat(settings): move voice dictation prefs into their own Dictation tab
The per-user voice settings (model, polish level, steering words) move out of the AI settings tab into a new 'Dictation' tab in user settings, mirroring the workspace Dictation tab. The AI tab keeps scratchpad instructions + encrypted scratchpads. Extracts a DictationSettings component; adds the 'dictation' tab to SETTINGS_TAB_OPTIONS, the tab config (with its own keywords), the dialog, and — automatically — the command palette.
